+++ api/routes/api.php
@@ -14,6 +14,7 @@ use App\Http\Controllers\Api\User\ResetPasswordController;
 use App\Http\Controllers\Api\Staff\SendOfferController;
 use App\Http\Controllers\Api\Staff\UserSearchController;
 use App\Http\Controllers\Api\User\Friku\FrikuJobsController;
+use App\Http\Controllers\Api\User\JobSearchesController;
 use App\Models\FrikuJoboffer;
 
 /*
@@ -43,7 +44,7 @@ Route::prefix('user')->group(function () {
     //Route::prefix('joboffer')内のグループにはjobofferがURIに付きます。
     Route::prefix('/joboffer')->group(function () {
         Route::get('/conditions', [JobsController::class, 'getConditions'])->name('joboffer.conditions');
-        Route::get('/search', [JobsController::class, 'searchJobOffers'])->name('joboffer.search');
+        Route::get('/search', [JobSearchesController::class, 'searchJobOffers'])->name('joboffer.search');
         Route::get('/{corporationJoboffer}', [JobsController::class, 'showJoboffer'])->name('pickup.show');
 
     });
